+++ TDLib.Api/Functions/GetBasicGroup.cs
@@ -31,14 +31,14 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("basic_group_id")]
-            public int BasicGroupId { get; set; }
+            public long BasicGroupId { get; set; }
         }
 
         /// <summary>
         /// Returns information about a basic group by its identifier. This is an offline request if the current user is not a bot
         /// </summary>
         public static Task<BasicGroup> GetBasicGroupAsync(
-            this Client client, int basicGroupId = default)
+            this Client client, long basicGroupId = default)
         {
             return client.ExecuteAsync(new GetBasicGroup
             {
