+++ TDLib.Api/Functions/SendBotStartMessage.cs
@@ -31,7 +31,7 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("bot_user_id")]
-            public int BotUserId { get; set; }
+            public long BotUserId { get; set; }
 
             /// <summary>
             /// Identifier of the target chat
@@ -52,7 +52,7 @@ namespace TdLib
         /// Invites a bot to a chat (if it is not yet a member) and sends it the /start command. Bots can't be invited to a private chat other than the chat with the bot. Bots can't be invited to channels (although they can be added as admins) and secret chats. Returns the sent message
         /// </summary>
         public static Task<Message> SendBotStartMessageAsync(
-            this Client client, int botUserId = default, long chatId = default, string parameter = default)
+            this Client client, long botUserId = default, long chatId = default, string parameter = default)
         {
             return client.ExecuteAsync(new SendBotStartMessage
             {
