+++ TDLib.Api/Functions/SetInlineGameScore.cs
@@ -34,7 +34,7 @@ namespace TdLib
             public string InlineMessageId { get; set; }
 
             /// <summary>
-            /// True, if the message should be edited
+            /// True, if the message needs to be edited
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("edit_message")]
@@ -45,7 +45,7 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("user_id")]
-            public int UserId { get; set; }
+            public long UserId { get; set; }
 
             /// <summary>
             /// The new score
@@ -66,7 +66,7 @@ namespace TdLib
         /// Updates the game score of the specified user in a game; for bots only
         /// </summary>
         public static Task<Ok> SetInlineGameScoreAsync(
-            this Client client, string inlineMessageId = default, bool editMessage = default, int userId = default,
+            this Client client, string inlineMessageId = default, bool editMessage = default, long userId = default,
             int score = default, bool force = default)
         {
             return client.ExecuteAsync(new SetInlineGameScore
