+++ TDLib.Api/Functions/ToggleSupergroupIsAllHistoryAvailable.cs
@@ -10,7 +10,7 @@ namespace TdLib
     public static partial class TdApi
     {
         /// <summary>
-        /// Toggles whether the message history of a supergroup is available to new members; requires can_change_info rights
+        /// Toggles whether the message history of a supergroup is available to new members; requires can_change_info administrator right
         /// </summary>
         public class ToggleSupergroupIsAllHistoryAvailable : Function<Ok>
         {
@@ -31,7 +31,7 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("supergroup_id")]
-            public int SupergroupId { get; set; }
+            public long SupergroupId { get; set; }
 
             /// <summary>
             /// The new value of is_all_history_available
@@ -42,10 +42,10 @@ namespace TdLib
         }
 
         /// <summary>
-        /// Toggles whether the message history of a supergroup is available to new members; requires can_change_info rights
+        /// Toggles whether the message history of a supergroup is available to new members; requires can_change_info administrator right
         /// </summary>
         public static Task<Ok> ToggleSupergroupIsAllHistoryAvailableAsync(
-            this Client client, int supergroupId = default, bool isAllHistoryAvailable = default)
+            this Client client, long supergroupId = default, bool isAllHistoryAvailable = default)
         {
             return client.ExecuteAsync(new ToggleSupergroupIsAllHistoryAvailable
             {
